+++ app/datatypes/user/actions.ts
@@ -1,12 +1,99 @@
 import {
-    UserActionTypes,
-    USER_RETREIVE_PUBLIC, USER_STORE_PUBLIC,
-    USER_RETREIVE_PRIVATE, USER_RETURN_PRIVATE, USER_RETURN_PRIVATE_FAILED,
-    USER_REQUEST_PERMISSIONS, USER_SUCCESS_PERMISSIONS, USER_FAILURE_PERMISSIONS, USER_CONFIRM_PERMISSIONS, USER_SUBMIT_BALLOT,
-    USER_RESET_REQUEST_PERMISSIONS,
-    UserPermissions,
+    ElectionPermissionRequest,
 } from "./types";
 import { ElectionId, CandidateId } from "../elections";
+
+export const USER_RETREIVE_PUBLIC = 'USER_RETREIVE_PUBLIC';
+export const USER_STORE_PUBLIC = 'USER_STORE_PUBLIC';
+
+export const USER_RETREIVE_PRIVATE = 'USER_RETREIVE_PRIVATE';
+export const USER_RETURN_PRIVATE = 'USER_RETURN_PRIVATE';
+export const USER_RETURN_PRIVATE_FAILED = 'USER_RETURN_PRIVATE_FAILED';
+
+export const USER_REQUEST_PERMISSIONS = 'USER_REQUEST_PERMISSIONS';
+export const USER_SUCCESS_PERMISSIONS = 'USER_SUCCESS_PERMISSIONS';
+export const USER_FAILURE_PERMISSIONS = 'USER_FAILURE_PERMISSIONS';
+
+export const USER_RESET_REQUEST_PERMISSIONS = 'USER_RESET_REQUEST_PERMISSIONS';
+
+export const USER_CONFIRM_PERMISSION = 'USER_CONFIRM_PERMISSION';
+
+export const USER_SUBMIT_BALLOT  = 'USER_SUBMIT_BALLOT';
+export const USER_SUCCESS_BALLOT = 'USER_SUCCESS_BALLOT';
+export const USER_FAILURE_BALLOT = 'USER_FAILURE_BALLOT';
+
+export interface UserRetreivePublicAction {
+    type: typeof USER_RETREIVE_PUBLIC;
+}
+
+export interface UserStorePublicAction {
+    type: typeof USER_STORE_PUBLIC;
+    payload: string;
+}
+
+export interface UserRetreivePrivateAction {
+    type: typeof USER_RETREIVE_PRIVATE;
+    payload: string;
+}
+
+export interface UserReturnPrivateAction {
+    type: typeof USER_RETURN_PRIVATE;
+    payload: string;
+}
+
+export interface UserReturnPrivateFailedAction {
+    type: typeof USER_RETURN_PRIVATE_FAILED;
+}
+
+export interface UserRequestPermissionsAction {
+  type: typeof USER_REQUEST_PERMISSIONS;
+  payload: ElectionPermissionRequest;
+}
+
+export interface UserSuccessPermissionsAction {
+  type: typeof USER_SUCCESS_PERMISSIONS;
+}
+
+export interface UserFailurePermissionsAction {
+  type: typeof USER_FAILURE_PERMISSIONS;
+}
+
+export interface UserResetRequestPermissionsAction {
+  type: typeof USER_RESET_REQUEST_PERMISSIONS;
+  payload: ElectionId;
+}
+
+export interface UserConfirmPermissionAction {
+  type: typeof USER_CONFIRM_PERMISSION;
+  payload: ElectionId;
+}
+
+export interface UserSubmitBallotAction {
+  type: typeof USER_SUBMIT_BALLOT;
+  payload: {electionId: ElectionId, receivers: {[key: string]: CandidateId} };
+}
+
+export interface UserSuccessBallotAction {
+  type: typeof USER_SUCCESS_BALLOT;
+}
+
+export interface UserFailureBallotAction {
+  type: typeof USER_FAILURE_BALLOT;
+}
+
+export type UserActionTypes = UserRetreivePublicAction
+                            | UserStorePublicAction
+                            | UserRetreivePrivateAction
+                            | UserReturnPrivateAction
+                            | UserReturnPrivateFailedAction
+                            | UserRequestPermissionsAction
+                            | UserSuccessPermissionsAction
+                            | UserFailurePermissionsAction
+                            | UserResetRequestPermissionsAction
+                            | UserConfirmPermissionAction
+                            | UserSubmitBallotAction
+                            | UserSuccessBallotAction
+                            | UserFailureBallotAction;
 
 /** retreive public key from the redux store */
 export function retreivePublicKey(): UserActionTypes {
@@ -65,8 +152,8 @@ export function permissionRequestFailure(): UserActionTypes {
 }
 
 /** Request of permissions for election has been confirmed */
-export function permissionRequestConfirmed(userPermissions: UserPermissions): UserActionTypes {
-    return { type: USER_CONFIRM_PERMISSIONS, payload: userPermissions };
+export function permissionRequestConfirmed(electionId: ElectionId): UserActionTypes {
+    return { type: USER_CONFIRM_PERMISSION, payload: electionId };
 }
 
 /** Submit ballot for election */
