+++ app/datatypes/user/reducer.ts
@@ -1,8 +1,8 @@
 import {
     USER_SUBMIT_BALLOT, USER_SUCCESS_BALLOT, USER_FAILURE_BALLOT,
     UserActionTypes, UserState, initialUserState, USER_STORE_PUBLIC,
-    USER_REQUEST_PERMISSIONS, USER_SUCCESS_PERMISSIONS, USER_FAILURE_PERMISSIONS, USER_RESET_REQUEST_PERMISSIONS, USER_CONFIRM_PERMISSIONS,
-} from "./types";
+    USER_REQUEST_PERMISSIONS, USER_SUCCESS_PERMISSIONS, USER_FAILURE_PERMISSIONS, USER_RESET_REQUEST_PERMISSIONS, USER_CONFIRM_PERMISSION,
+} from ".";
 
 /** reducer for user */
 export function user(
@@ -28,9 +28,12 @@ export function user(
             return { ...state, permissionRequestStatus: "Success" };
         case USER_FAILURE_PERMISSIONS:
             return { ...state, permissionRequestStatus: "Failed" };
-        case USER_CONFIRM_PERMISSIONS:
+        case USER_CONFIRM_PERMISSION:
             return { ...state,
-                permissions: action.payload,
+                permissions: { ...state.permissions,
+                    canVote: [ ...state.permissions.canVote.filter(v => v !== action.payload), action.payload ],
+                },
+                activePermissionRequest: state.activePermissionRequest.filter(v => v !== action.payload),
             };
         case USER_SUBMIT_BALLOT:
             return { ...state, ballotSubmissionStatus: "Fetching" };
