+++ package.json
@@ -20,6 +20,7 @@
     "axios": "^0.19.0",
     "buffer": "^5.4.3",
     "hash.js": "^1.1.7",
+    "isomorphic-ws": "^4.0.1",
     "mem": "^4.3.0",
     "node-fetch": "^2.6.0",
     "react": "^16.11.0",
@@ -39,7 +40,8 @@
     "react-scripts": "^3.3.0",
     "redux": "^4.0.4",
     "redux-saga": "^1.1.3",
-    "scrypt-js": "^3.0.0"
+    "scrypt-js": "^3.0.0",
+    "ws": "^7.2.1"
   },
   "devDependencies": {
     "@react-native-community/eslint-config": "^0.0.5",
